Add files via upload
The contract is an ERC721 token smart contract that inherits from the ERC721Enumerable and Ownable contracts. It has the following functions:

constructor: It initializes the contract with a name, symbol, and base URI. It also mints the first token to the contract deployer.

mint: This function allows users to mint tokens by sending ether to the contract. It requires that the contract is not paused, the mint amount is greater than zero, the mint amount is less than or equal to the maximum mint amount, and the total supply plus the mint amount is less than or equal to the maximum supply. If the sender is not the owner, it requires that the sender's address is whitelisted or the value sent is greater than or equal to the cost multiplied by the mint amount. It mints the tokens to the specified address.

count: It returns the total number of tokens that have been minted.

walletOfOwner: This function returns an array of token IDs owned by a specified address.

tokenURI: This function returns the URI of a given token. If the isRevealed flag is true, it returns the token URI; otherwise, it returns the preRevealedURI.

setPreRevealedURI: This function allows the contract owner to set the preRevealedURI.

revealAndSetBaseURI: This function allows the contract owner to set the baseURI and set the isRevealed flag to true.

setCost: This function allows the contract owner to set the cost of minting tokens.

setmaxMintAmount: This function allows the contract owner to set the maximum number of tokens that can be minted at once.

setBaseURI: This function allows the contract owner to set the base URI of the token.

setBaseExtension: This function allows the contract owner to set the base extension of the token.

pause: This function allows the contract owner to pause or unpause the contract.

whitelistUser: This function allows the contract owner to whitelist an address.

removeWhitelistUser: This function allows the contract owner to remove an address from the whitelist.

withdraw: This function allows the contract owner to withdraw the contract's balance.